move capture server config into a function
